01_python_core/084_Sandbox/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
pydantic==2.13.1
pydantic_core==2.46.1
python-dotenv==1.2.2
python-multipart==0.0.26
requests==2.33.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
